+++ priv/storybook/_root.index.exs
@@ -5,5 +5,5 @@ defmodule Storybook.Root do
   use PhoenixStorybook.Index
 
   def folder_icon, do: {:fa, "book-open", :light, "lsb-mr-1"}
-  def folder_name, do: "Storybook"
+  def folder_name, do: "Doggo"
 end